serialization structure, project dir macros...
new character constructor, modified some character fields types, passing character ptr to first sequence functions...

+++ Weapon.cpp
@@ -34,9 +34,9 @@ bool Weapon::isBroken() const {
 string Weapon::getName() const {
 	return this->_name;
 }
-float Weapon::getDurability() const {
+int Weapon::getDurability() const {
 	return this->_durability;
 }
-float Weapon::getDamage() const {
+int Weapon::getDamage() const {
 	return this->_damage;
 }
